# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface WaitlistEntry {
  id: number;
  /** @pattern ^[^\s@]+@[^\s@]+\.[^\s@]+$ */
  email: string;
  twitterUsername: string;
  xFollowed: boolean;
  telegramJoined: boolean;
  createdAt: string;
}

export interface WaitlistEntryInput {
  /** @pattern ^[^\s@]+@[^\s@]+\.[^\s@]+$ */
  email: string;
  /** @minLength 2 */
  twitterUsername: string;
  xFollowed: boolean;
  telegramJoined: boolean;
}

export interface WaitlistSummary {
  total: number;
  joinedToday: number;
  xCompleted: number;
  telegramCompleted: number;
}

export interface ErrorResponse {
  error: string;
}

export type ListWaitlistEntriesParams = {
search?: string;
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
};