# lib/api-zod/src/generated/types/waitlistEntry.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WaitlistEntry {
  id: number;
  /** @pattern ^[^\s@]+@[^\s@]+\.[^\s@]+$ */
  email: string;
  twitterUsername: string;
  xFollowed: boolean;
  telegramJoined: boolean;
  createdAt: Date;
}