# lib/api-zod/src/generated/types/waitlistEntryInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WaitlistEntryInput {
  /** @pattern ^[^\s@]+@[^\s@]+\.[^\s@]+$ */
  email: string;
  /** @minLength 2 */
  twitterUsername: string;
  xFollowed: boolean;
  telegramJoined: boolean;
}