# lib/api-zod/src/generated/types/waitlistSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WaitlistSummary {
  total: number;
  joinedToday: number;
  xCompleted: number;
  telegramCompleted: number;
}